[FEATURE] Support pecl-memcached in MemcachedBackend
The MemcachedBackend now also supports the pecl-memcached
module. The caching backend detects which modules are installed
and uses "memcache" over "memcached" to avoid being a breaking
change. If both modules are installed, an integrator can choose the
PECL module by setting the new ``peclModule`` option.

This feature was originally introduced with TYPO3 v8.0. Due to
compatibility reasons with PHP 7 this feature gets backported as the
memcache module is not available anymore.

Resolves: #83569
Related: #69794
Releases: 7.6
Change-Id: Idd4e85cf8ec71d47217b63dfe55b1231107c7b82
Reviewed-on: https://review.typo3.org/55384
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>