[BUGFIX] Timeout processing of media URL to avoid race condition
The new "Add media by URL" button opens a modal with a textfield
for an URL. If processing the URL fails, an error modal is displayed.
However, because of a race condition the error dialog is not closable.
Add a timeout of 200ms before processing the URL to workaround the
race condition.

Resolves: #69868
Related: #61799
Releases: master
Change-Id: I7a8f1450b73b4923a1ae9c2e0346e2571475b92e
Reviewed-on: http://review.typo3.org/43369
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>