[TASK] Un-deprecation log entries for ext:belog
The usage of __wakeup() in Constraint Model
(by extending AbstractDomainModel) causes deprecation log entries.

The core must not cause any deprecation log entries, so as temporary
fix an empty __wakeup() function was added. The Constraint Model is
subject to refactoring.

With #62925 and the properties manualDateStart and manualDateStop
have been marked as deprecated with the assumption they are not in
use. But they are, and so they are restored as fully functional
properties.

Change-Id: Iff23f10617700751991026706f7d6284de93e469
Releases: master
Related: #62925
Resolves: #70000
Reviewed-on: http://review.typo3.org/43457
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>