[BUGFIX] Do not call non static methods statically
ClassInformationGenerator used to be a class with static methods.
Remove all static calls to methods which are not static any more.

Also make the property non static, so that we do not have to reset
the state if we do generate the class info twice in one request.

Resolves: #70057
Releases: master
Change-Id: I1323d40b1181676001f08c8a791c1f312308389c
Reviewed-on: http://review.typo3.org/43509
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>