[BUGFIX] Fix validation for objects of the same type
The patch http://review.typo3.org/40693 introduced a regression by
replacing a count() check on a countable object with a !empty() check,
which lead to failing validations for objects of the same type.

Since getBaseValidatorConjunction() returns either an object or NULL
check for NULL instead and make the countable call explicit.

getBaseValidatorConjunction() doesn't have a second parameter in the
backport, thus remove that parameter.

Resolves: #70017
Related: #67786
Releases: master
Change-Id: I141ecb0b14f6dcebc2908a27009abc0c3f3ccf75
Reviewed-on: http://review.typo3.org/43564
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>