Revert "[!!!][BUGFIX] Skip cache hash for URIs to non-cacheable actions"
This reverts commit 3ab6f2a08283041d52a05db3e26817c66024da88.

This change caused trouble, as it was not possible any more to explicitly enable
cHash usage. In situations where some arguments are required for a cHash, this
will lead to a broken page (404) when the cHash is skipped.

As the possible gain of not having the cHash is low but the possible impact is high
and a potential fix is breaking or at least fragile as well, this feature is reverted again.

Resolves: #67097
Reverts: #60272
Releases: master
Change-Id: Ia14947f8eb6b6125b0402ae7e046c78a1597aeae
Reviewed-on: http://review.typo3.org/43688
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>