[BUGFIX] BE: correctly resolve default values for MM relations
With this default values for MM relations are
again retrieved correctly.
This bug was introduced with the "FormEngine" changes
because the values for MM tables are retrieved via
RelationHandler now and always use the MM table.
In order for this to work for default values the
MM table needs to be ignored.

Resolves: #70275
Releases: master
Change-Id: Ice3660c3e2d3c172034f308a642a92d4cbe2f618
Reviewed-on: https://review.typo3.org/43727
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>