[BUGFIX] Fluid form fields only convert identifier objects to identifier
With the change from getValue to getValueAttribute Fluid now converts
all objects to identifier objects which is wrong because there are many
objects that are valid without identifiers.
For instance, ObjectStorage in multiple select field.
Or a object that implements __toString in TextfieldViewHelper.

The change was introduced with commit https://review.typo3.org/42298/

Change-Id: I666112feeb05381553e8a22881773689f84491b7
Resolves: #70302
Releases: master
Reviewed-on: http://review.typo3.org/43735
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>