[!!!][TASK] Change EXT:form wizard to use AJAX routes
The ExtJS wizard of EXT:form in the backend is using
an old "module" to load and to save the data from the wizard,
but is misusing this as AJAX responses.

The patch uses PSR-7 Ajax Routes to load the content,
and also removes the now obsolete WizardView classes.

Resolves: #70574
Releases: master
Change-Id: I714b514ea75f618db58c18f3e0ac1235cfd40f7e
Reviewed-on: https://review.typo3.org/43881
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>