[BUGFIX] Fix Liskov substitution violation in PackageManager
TYPO3's `PackageManager` class changes the signature of a method
defined in it's parent class. This causes warnings when PHP's strict
mode is enabled. Fix this in the same way, we established compatibility
with the Flow core bootstrap class and establish an alias for the
Flow core class loader.

This brings strict free TYPO3 front end rendering without having
to change the method signature.

When doing so, we can also remove the Flow class loader class from
the class map and the file system.

Change-Id: I29ce35339823126bbba0c951da11122e16b4e052
Fixes: #69476
Releases: 6.2
Reviewed-on: http://review.typo3.org/42983
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Martin Helmich <m.helmich@mittwald.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>