[TASK] Update static composer class loading
The committed composer class loading information has not been
updated for a while. Do so, to add classes that have been added
meanwhile to required TYPO3 extensions to the map and update
class loader code to be current.

Also adapt the .gitignore file to be correct and add a script
to automate this process (so that it can theoretically be done
before a release)

Resolves: #70465
Releases: 6.2
Change-Id: Ife3a9b47a6fbd93e4d10c6e693b597f42a338bc9
Reviewed-on: http://review.typo3.org/43887
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>