[BUGFIX] Repair System Information Panel extendibility
Contrary to the documentation given for extending the System Information
Panel via SignalSlot only a single item can be added for each slot due to
improper use of the SignalSlot pattern. The return value of both signals is
assumed to be only one item. This patch passes the Controller as an argument
and adds two methods to add items.

The two signals were marked as internal, so this is not considered
a breaking patch.

Resolves: #70567
Releases: master, 7.6
Change-Id: I268296860ac2f0eb2344f84604a096999bdccac0
Reviewed-on: https://review.typo3.org/43973
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
(cherry picked from commit b9d5cf143fbdf373177203d9d4e7b792d43295dc)
Reviewed-on: https://review.typo3.org/47487