[BUGFIX] BE DatePicker allow manual date & time input
FormEngineValidation currently destroys the input date from datepicker.
In order for the date & times to be handled correctly.
We need to parse the values correctly
(some bugs during refactoring to FormEngineValidation)
and afterwards reformat the value so that when
DatePicker resolves the value it is in the correct format.

Change-Id: I7d40cd2243977999537b76176b59d58829c0ef6b
Resolves: #70161
Resolves: #68907
Releases: master
Reviewed-on: http://review.typo3.org/43614
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>