[BUGFIX] Do not crop inside known html tags
This change adds missing tags to the regex used to match html tags.
Furthermore, scripting tags are added, so that they will not be counted
nor split apart.

This list of html tags was extracted from the current W3C html5
standard recommendation at [1] and merged with tags from the current
code that are not (or no longer) in the W3C standard recommendation
(e.g: TYPO3's internal <link> tag).

---
[1]: http://www.w3.org/TR/html5/semantics.html#semantics

Resolves: #69263
Releases: master, 6.2
Change-Id: Id58dcf3470ed06256b9b8f744dd13891757faef1
Reviewed-on: http://review.typo3.org/43995
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>