[BUGFIX] Flexforms were not considered for FormEngine data preparation
With #70504 a selective collection of fields to be processed has been
introduced in FormEngines data providing mechanism. This collection
did not respect fields added by extensions, like the flexform field,
so no flexform was displayed in the end.
Those fields are now added to the process again and flexforms are
displayed as requested.

Change-Id: I49f9137f44588df735d30331465c230f1daee96b
Resolves: #70626
Releases: master
Reviewed-on: http://review.typo3.org/44012
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Stefan Froemken <froemken@gmail.com>
Tested-by: Stefan Froemken <froemken@gmail.com>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>