[TASK] Add new property displayName to ShortcutButton
Actually the ShortcutButton shows the module name as label for a shortcut.
For modules with different actions inside the module this results with
the same label for all actions of a module.

This patch add a new property displayName to the ShortcutButton.

With this patch the label in the bookmark list can be set when creating
the ShortcutButton using the new property displayName.
If displayName is set it is used as label for the bookmark,
if displayName is not set (empty) the ShortcutButton works as before.

Resolves: #70698
Releases: master
Change-Id: Ie650d08793f323581ebee52af59c32a8a85b08c9
Reviewed-on: https://review.typo3.org/44065
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>