[BUGFIX] FAL relations duplicated when saving in workspaces
In a workspace, when a record is saved which contains an IRRE-managed
relation in a flexform field, the relation is duplicated. The reason is
that the workspace information is lost along the way when processing the
flexform content.

This commit introduces an additional method parameter to the relevant
flexform processing callbacks.

Additionally this fixes an error in the Flexform XML fixture, which
(apparently) made the test go into an endless loop.

Work time for this patch was sponsored by Matthias Krappitz (aemka.de),
the time for developing the test was sponsored by me.

Change-Id: If6d71cf966408fcf976365fc01e6a2529bf1902d
Releases: master, 7.6, 6.2
Resolves: #70962
Change-Id: I1cc92768608aeee419290d75dae5c34b7560e379
Reviewed-on: https://review.typo3.org/50313
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Wolf <andreas.wolf@typo3.org>
Reviewed-by: Markus Sommer <markussom@posteo.de>
Tested-by: Markus Sommer <markussom@posteo.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>