[BUGFIX] Add class alias of class loader to alias map
Instead of hardcoding the class_alias call to alias the Flow
class loader, we must use the class alias API,
which in turn triggers the integrated composer
class loader to properly establish the alias.

This avoids a fatal error in the install tool clear cache action
which calls Bootstrap::initializeClassLoader() twice.

Resolves: #71057
Related: #69476
Releases: 6.2
Change-Id: I83310f9f9668c2ce0cdd07497036b406e94decc7
Reviewed-on: https://review.typo3.org/44302
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>