[BUGFIX] Empty grid list in frontend workspace preview
The grid list of the frontend workspace preview stayed empty due to
several JavaScript issues. First one is a CORS thingy with local
document.domain values being set and for the list IFRAME being set
too late (and thus running into CORS error). Besides that the tabs
component is only available in the backend workspace module and
leads to another error if expected and used in the frontend preview.

Resolves: #66169
Releases: master, 6.2
Change-Id: I5bf3dd7760c9382527d85e6e7e8e0ad5dca81c05
Reviewed-on: https://review.typo3.org/44307
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>