[BUGFIX] Add reload parameter for EM List Utility
If the extension manager list utility "getAvailableExtensions()" is called
and then a new package is added at the same request, it is not possible
to regenerate the list of availableExtensions as this is cached in a run-time
member variable inside the list utility.

A new method "reloadAvailableExtensions" is added in order to
rebuild the list of available extensions.

Resolves: #71071
Releases: master
Change-Id: Idab208c27f8905e3770fa1845a805cf3f5c3e8ac
Reviewed-on: https://review.typo3.org/44318
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>