[BUGFIX] Make translating inline elements possible
If a translated parent record has inline childs its
now possible to translate new childs of the translation
parent of parent record.

For example if a tt_content element has new images
added to the default language its now possible to add
the translation for the image reference in the
translated content element.

Resolves: #70880
Releases: master
Change-Id: I6c19aaa9238232d0076509ce92fe84552fd69c18
Reviewed-on: https://review.typo3.org/44338
Reviewed-by: Benjamin Kott <info@bk2k.info>
Tested-by: Benjamin Kott <info@bk2k.info>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>