[BUGFIX] FormEngine: Pass correct selected node count to ExtJS Tree
The selected node count for the ExtJS Tree wasn't calculated using the
raw database field value instead of the parsed tree configuration. In
case of empty database values this resulted in a wrong count of selected
nodes, disabling the selection of nodes in the case of maxitems=1 in the
tree configuration.

Resolves: #70306
Resolves: #70630
Related: #69836
Releases: master
Change-Id: Ib04207c60b7a3473eaac619b6f2f4091aa5b938f
Reviewed-on: https://review.typo3.org/44356
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>