[BUGFIX] Suggest wizard does not work on new records
The suggest wizard gets it's parameters via a POST json
body. Inside there is the current record uid given,
which follows the scheme "NEW{random number}". For whatever
reason the wizard casts this uid to an integer which results
in a 0. The following check for a numeric uid fails in the
case of a new record, as it is always numeric because of the
cast.

This patch removes the cast. No further sanitizing is necessary
as the uid is casted in every further use case where appropriate.

Resolves: #71056
Releases: master

Change-Id: I1f20fbc32ab406e69596ad8bf7a1e65231b9f1d6
Reviewed-on: https://review.typo3.org/44412
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Gernot Schulmeister <gernotschulmeister@gmx.at>
Tested-by: Gernot Schulmeister <gernotschulmeister@gmx.at>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>