[BUGFIX] EXT:form - HIDDEN element cannot be accessed in form wizard
After creating a HIDDEN element with the help of the form wizard one is
not able to edit the element afterwards. This is due to the fact that
the element is hidden which is kind of correct but does not help the
editor to work target orientied.

This patch adds a dummy element to the HIDDEN element. The dummy
element contains the element's name and has a dotted border.
That way the editor can edit and move a HIDDEN element.

Resolves: #70637
Releases: master, 6.2
Change-Id: I933df4f7a4b01d9246f6bd0981179497c103882b
Reviewed-on: https://review.typo3.org/44434
Reviewed-by: Bjoern Jacob <bjoern.jacob@tritum.de>
Tested-by: Bjoern Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>