[BUGFIX] Ignore placeholders when copying records
Copying records would make placeholders for draft elements visible in the
copy. This is due to the fact that t3ver_state information is removed upon
copy and the records itself are not ignored either. To circumvent that, now
only relevant records are selected for each scope, being live workspace or
any draft workspace. This also means, that new placeholders are copied if
the process has been initiated on a real draft workspace, which does not
happen if the duplication process was triggered in the live workspace.

Resolves: #36946
Resolves: #42075
Releases: master, 6.2
Change-Id: I0952d1a46d1cc02e0ec97f7cbbc9abc700e00c07
Reviewed-on: https://review.typo3.org/45071
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>