[BUGFIX] Adodb PHP7 Redefinition of parameter $owner
Adodb contains two methods that declare parameter $owner twice.
The PHP 7 linter fatals with this. A fix was merged to upstream
with commit ae85debec8 on https://github.com/ADOdb/ADOdb/ but not
released yet. The patch picks this part from the patch.

In result only linter warnings are left for old constructors in
openid and adodb, so travis-ci is green on PHP7 and can be removed
from "allowed_failures".

Resolves: #71244
Releases: master
Change-Id: I3b2fd4c7f9efcbe5bb0cf70df84dfe8d219dccd0
Reviewed-on: https://review.typo3.org/44477
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>