[FEATURE] Add FlashMessage support to ModuleTemplate API
Flash messages have different queues, depending on in which context they
are enqueued. The FlashMessageService defaults the queue to
``core.template.flashMessages``, Extbase defaults the queue to
``extbase.flashmessages.$randomPluginQueue``.

Add support for flash messages to ModuleTemplate which makes it possible
to enforce a queue identifier which is rendered in a global scope.

Resolves: #71251
Releases: master
Change-Id: I7e0a0088f8020994dbff162ed5f3e53b6ceeedce
Reviewed-on: https://review.typo3.org/44480
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>