[BUGFIX] Empty array not encoded as empty XML tag
When encoding an empty array, an empty line with indentation was
created. If the <el> tag had a proper type="array" attribute, it could
still be decoded to an array again. If not, it was decoded to a string,
leading to errors e.g. when trying to traverse it.

The fix is to include an empty element now, which will be decoded to
an empty array again then.

Change-Id: Ia78adc1e7e81a943cdbc94f3cfa19b36442c430f
Resolves: #71394
Releases: 6.2, master
Reviewed-on: https://review.typo3.org/44601
Reviewed-by: Daniel Lorenz <forge@extco.de>
Tested-by: Daniel Lorenz <forge@extco.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>