[BUGFIX] BackendUtility: Skip rendering thumbnails for anything but files
If the stored value in a group field with type file/file_reference is a
folder the thumbCode() method skips rendering a thumbnail for that
resource. This avoids fatal error as the interface of e.g. a Folder is
different to that of a File, missing the ::isMissing() method.

Resolves: #66045
Releases: master, 6.2
Change-Id: I29b75dec0173cc2423583a9cf42ff56ca81ce7d3
Reviewed-on: https://review.typo3.org/44709
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>