[BUGFIX] Handling of uploaded files not within open_basedir
When the "upload_tmp_dir" folder is not within the "open_basedir"
an Exception will get thrown when uploading files in the fileadmin.
This patch silences open_basedir() warnings occuring in the file_exists()
check and adds detection of uploaded files as the handling of these with
move_uploaded_file() is not influenced by the open_basedir setting.

Resolves: #69356
Releases: master, 6.2
Change-Id: I649373e8910c2d30f73cbb0e3b56476196c151fd
Reviewed-on: https://review.typo3.org/44710
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>