[BUGFIX] Update tstamp on writeForeignField if any foreign value has changed. (IRRE)
This patch solves the problem where the timestamp in the column tstamp
was not updated when the sorting of the inline list (IRRE) would change.
It also updates the tstamp if one of the following fields change:
foreign_field, foreign_table_field, symmetric_field or one of the
foreign_match_fields.

Resolves: #71138
Releases: master
Change-Id: I3c9143dfd7bc2dca3ce5eb1dbe1f8e71561ebf66
Reviewed-on: https://review.typo3.org/44617
Reviewed-by: Oliver Eglseder <typo3org@vxvr.de>
Tested-by: Oliver Eglseder <typo3org@vxvr.de>
Reviewed-by: Markus Sommer <markussom@posteo.de>
Tested-by: Markus Sommer <markussom@posteo.de>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>