[TASK] Remove Ter.json template in extension manager
The Ter.json template is needed to reload the extension table list
after the extension information was reloaded. As this request is not
useful (e.g. any search parameter stays in the input field, but the
request doesn't filter anything), the template should be removed and
the whole page should be reloaded afterwards.

Resolves: #71669
Releases: master, 6.2
Change-Id: I64158edff282540e7cc9f2597c9579607834aa55
Reviewed-on: https://review.typo3.org/44778
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>