[BUGFIX] Allow reply-to mail header in EXT:form
In EXT:form it was not possible to set a reply-to header in the email
and so the entered mail address was used as "from" which could cause SPF
failures.

The reply-to was there in the old mailform (e.g. TYPO3 4.5) but not in
the new extension which is the successor. That is why this issue handled
as bugfix.

Resolves: #69395
Related: #68771
Releases: 6.2

Change-Id: I1e7b77d3196c4a921d2902f212d03815bb3e0eb9
Reviewed-on: https://review.typo3.org/43661
Reviewed-by: Bjoern Jacob <bjoern.jacob@tritum.de>
Tested-by: Bjoern Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>