[!!!][TASK] Remove class loading fallback in cObject
Remove cObj->isClassAvailable() as this class was
introduced in TYPO3 4.3 before autoloading was there.

Now, with TYPO3 CMS 7, autoloading is completely done
via spl_autoload_register and class_exists() does the
autoloading automatically, both in composer-mode and
non-composer-mode.

The method is therefore removed as the fallback is
pre-historic and a simple class_exists() check is added.

Resolves: #72022
Releases: master
Change-Id: I92eef8dee6277bcec63403dcf04e12955cb669f2
Reviewed-on: https://review.typo3.org/45085
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>