[BUGFIX] Let FormEngine Slider wizard save data again
Due to a change in FormEngine slider rewrite, the parameters
do not get handed over to the TBE_EDITOR update signal,
thus any changes on the slider do not reflect the actual field
that is sent to the server again.

Using apply() correctly as well as a document.ready()
statement makes this functionality work again.

Resolves: #71977
Releases: master
Change-Id: Ib033015d0c3869da487e9163ceca303bd1b436c8
Reviewed-on: https://review.typo3.org/45145
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>