[BUGFIX] Hide report notice in about module for editors
Editors are not allowed to see the reports module, therefore the link
'Please check the status report for more information' does not work.
Furthermore those users can't change anything which is reported, so
don't show this information at all.

Displaying less warnings which can not be changed improve the usability.

Change-Id: I552b34f4bc1e249c51abc166699597b1678a61f9
Resolves: #72084
Releases: master
Reviewed-on: https://review.typo3.org/45168
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>