[BUGFIX] Change DB field folder for sys_file_collection to allow NULL
Change the database schema to not require a value in the optional folder
column of the sys_file_collection table. Data for this field is only
provided when the type of collection created in the backend is based on
a folder, leading to errors in MySQL strict mode.

Resolves: #72305
Releases: master, 7.6
Change-Id: I35520061d5194db8328a43fe6e0d1887972ab889
Reviewed-on: https://review.typo3.org/45346
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>