[BUGFIX] FormEngine: Fix expanding of inline records in FlexForms
This commit provides the proper configuration information to the
compiler and render so that inline records in FlexForms can be
successfully deleted after being expanded from a collapsed state.

This change also fixes the problem that inline records within a
flexform displayed the default field palette due to missing
configuration information.

Resolves: #72294
Releases: master, 7.6
Change-Id: I330ab93a6127e1f84ed21ff6984721fb655d9136
Reviewed-on: https://review.typo3.org/45364
Reviewed-by: Armin Ruediger Vieweg <armin@v.ieweg.de>
Tested-by: Armin Ruediger Vieweg <armin@v.ieweg.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>