[BUGFIX] Handle empty pid list in FrontendConfigurationManager
The result of getTreeList() called in overrideStoragePidIfStartingPointIsSet()
may be an empty string, which is appended to the $list array.
Empty values should not be added to the array, since it will cause an invalid
list of pages later on.

Add a check to avoid empty values in $list.

Resolves: #72414
Releases: master, 7.6
Change-Id: Id1309a44f46dcd84ab9923da1be6873bdc38cb8d
Reviewed-on: https://review.typo3.org/45430
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>