[TASK] Expose identifier and path of functional test instance
Identifier and path of a functional test instance is created during
bootstrapping the testcase. However, if one needs to define particular
path settings to the initialization phase, this will end up in being
a chicken-or-the-egg problem.

That's why the mentioned two parts are exposed as static functions
and wrapped by the functional test base class.

Resolves: #72450
Releases: master, 7.6, 6.2
Change-Id: I111768133456974010d49b02225e41f9b74dbcff
Reviewed-on: https://review.typo3.org/45486
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>