[!!!][TASK] Remove RTE transformation ts_strip
The conversion to strip certain all HTML tags out
of the RTE parser (ts_strip) on the way from the RTE
to the database is removed.

The hardcoded static list is superseeded by the
HTMLparser logic and all TSconfig used since 4.5 anyway,
so this static logic is obsolete.

Resolves: #72661
Releases: master
Change-Id: I7db74c99101e39d37191601c5e77a5abc04e8e6e
Reviewed-on: https://review.typo3.org/45848
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>