[TASK] Allow access to visible content within /.well-known/ directory
Allow access to the visible content from within the `/.well-known/`
hidden directory. The access to all other hidden files and directories
(starting with a dot) is still blocked.

The /.well-known/ directory represents the standard (RFC 5785) path
prefix for "well-known locations", and therefore, access to its visible
content should not be blocked.

Resolves: #72712
Releases: master,7.6,6.2
Change-Id: I533d38a12da5cae59abed4fc00d597814d28fa04
Reviewed-on: https://review.typo3.org/45927
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>