[BUGFIX] Import skips files with non-existent target directory
When an import is taking place in which sys_file_storage records
get imported any files within this storage will not get imported
if the basePath of the sys_file_storage does not exist.

This patch displays an error message in such cases and prompts
the user to create the missing directory.

Resolves: #68791
Releases: master, 7.6
Change-Id: I7fb0f0cdf9b25b29b0a35781450020a59e0f03ab
Reviewed-on: https://review.typo3.org/45975
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>