[!!!][TASK] Remove RTE transformation option preserveTables
The RTE transformation option "preserveTables" that
allows to keep HTML table elements and their contents
when not in transformation mode "ts_css" (but just in"ts")
is removed.

Use ts_css instead which keeps the tables as they are completely.

If the RTE should be configured to remove HTML tables
completely this can be done via denyTags := addToList(table).

Resolves: #72783
Releases: master
Change-Id: I008e809a51ce38938adfff64cadc2686d7304f9a
Reviewed-on: https://review.typo3.org/46041
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>