[!!!][TASK] RTE: Always wrap RTE lines in <p> not <div> tags
The rtehtmlarea RTE shipped with the TYPO3 Core does not
work with the RTE options "useDIVasParagraphTagForRTE"
and "remapParagraphTag" which allowed the content sent to the
RTE to use <div> tags or no tags instead of the regular <p>
tags.

Using these options before resulted in having multiple <div>
tags added every time the contents of the RTE were saved
to the database.

Since the RTE does not support this functionality, nor does
it make any sense, it is removed.

Resolves: #72866
Releases: master
Change-Id: I684594dc5eea98e9f88ef69cb2edf592c3a67516
Reviewed-on: https://review.typo3.org/46130
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Thomas Schlumberger <thomas@b13.de>
Tested-by: Thomas Schlumberger <thomas@b13.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>