[!!!][TASK] EXT:form - Remove deprecated code
This patchset removes the compatibility layer of EXT:form. The layer
made sure that EXT:form acts almost like in TYPO3 6.2. This mainly
applies to the layout configuration and rendering.
In the former days integrators could use .layout settings on different
levels to change the output of the form elements in the frontend. The
whole process is not intuitive and not well documented.
Nowadyas integrators should use fluid templates to customize their
forms. With this patchset this is the only way to go.

Furthermore the SELECT, TEXTAREA and TEXTBLOCK elements are adjusted.
The automatic transformation of the .data and .content attribute is
removed. That way .text is the only valid attribute for adding a human
readable text to the above mentioned elements.
Using the old attribute causes no harm.
There will be neither an output nor an error message.

Resolves: #72861
Releases: master
Change-Id: Ibe35e68637b904c7db705d3c8fcf70b77dc8a1b1
Reviewed-on: https://review.typo3.org/45384
Reviewed-by: Bjoern Jacob <bjoern.jacob@tritum.de>
Tested-by: Bjoern Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>