[BUGFIX] SimpleFileBackend::setCacheDirectory returns valid path name
The comment for the method
\TYPO3\CMS\Core\Cache\Backend\SimpleFileBackend::setCacheDirectory()
states clearly, that it is possible to provide an absolute path as
cache directory. However, in the lower part of the function, it is
stated, that if ($cacheDirectory0 == '/') the documentRoot should be
set to '/'. That results in a returned path like '//...'.
This causes problems within
\TYPO3\CMS\Core\Utility\GeneralUtility::validPathStr(), because pathes
with '//' in it are always invalid.

Resolves: #72635
Releases: master, 7.6, 6.2
Change-Id: I30e7743ba9835c99382b7c04153f91e688ead5fc
Reviewed-on: https://review.typo3.org/46157
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jan Helke <typo3@helke.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>