[TASK] Add info about Apache version when using mod_filter
Since mod_filter is available since Apache 2.3.7 we need to check for
the apache version in the htaccess file.
Older versions of apache will work as well, even though they do not need
to check for the existence of mod_filter.

A comment is added to inform older Apache versions.

Resolves: #72886
Releases: master, 7.6, 6.2
Change-Id: Ia4905c992b52b2bd540ece0a1c1866aeacf6de85
Reviewed-on: https://review.typo3.org/46150
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>