[BUGFIX] feedit: Double encoding of labels
Translated labels containing UTF-8 characters were double-encoded
in frontend editing.

extGetLL() already runs htmlspecialchars() on labels, so we do not have
to do that again in FrontendEditPanel.
This fixes the titles of the admin panel buttons.

Also, since we can expect all pages to be UTF-8 now, we can drop the forced
conversion of UTF-8 characters to XML entities.
This fixes the confirmation popup texts.

Resolves: #63131
Releases: master, 7.6
Change-Id: I49faff20ff4da8434868534d566c0d3da85e30c6
Reviewed-on: https://review.typo3.org/46802
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Marco Huber <mail@marco-huber.de>
Tested-by: Marco Huber <mail@marco-huber.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>