[BUGFIX] Remove crawler initialization in indexed search
There is a leftover stub run from the global scope to ensure
that the hook uses the proper backend charset. This code
was never executed in the current version of crawler, but
still may be viable in some situations. Therefore it is added
to the constructor method of the actual Crawler Hook.

Resolves: #72998
Releases: master, 7.6
Change-Id: Ie091c2bbe7c3eafaf8ab33ac2065501d44c35bd8
Reviewed-on: https://review.typo3.org/46412
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>