[BUGFIX] Inline record in Flexform can attach new record
If you click the "create new element" button on an inline record within
a FlexForm field, the passed data information contains the id after
which the new element should be inserted. Currently this id isn't
respected while receiving the configuration from the given data. This
patch adds code to respect a possible given id and allows to fetch
configuration for that element as well.

Resolves: #73004
Releases: master, 7.6
Change-Id: I70263d3ff4a8641b1a8280574521e0671c34a333
Reviewed-on: https://review.typo3.org/47034
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>