[TASK] Have PackageManager only check for TYPO3 extensions
Currently the PackageManager class checks for TYPO3 extensions
and for directories within typo3/sysext/ typo3/ext/ or typo3conf/ext/
that contain a composer.json but no ext_emconf.php.

However, as the additional check, which only checks for the
existance of composer.json was part of the legacy functionality
from the old Flow backport and is not used anymore.

Resolves: #73019
Releases: master
Change-Id: I494030d6ee133b4db35387e20b326f9215edfc4e
Reviewed-on: https://review.typo3.org/46377
Reviewed-by: Daniel Maier <dani-maier@gmx.de>
Tested-by: Daniel Maier <dani-maier@gmx.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>