[TASK] Do not hide possible PHP warning from Extbase exception
The Extbase ClassInfoFactory may be unable to build information for a
given class if a PHP warning occurs while running in development context.

Instead of hiding the actual exception with a generic message
"Could not analyse class: VendorName\ExtensionKey\Some\Class", it should
include the message of the corresponding exception to give useful advice
while fixing the problem.

Change-Id: I7c4b5c13b034d9a9a0d24aab84863008fbd07598
Releases: master, 7.6
Resolves: #73053
Reviewed-on: https://review.typo3.org/46453
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Tested-by: Xavier Perseguers <xavier@typo3.org>