[BUGFIX] Reallow '0' as valid userFunc argument in TypoScript
With #47301 the parsing of userFunc in TypoScript has changed.
The change prevents '0' as valid argument.

This patch fix the '0' bug but also adds some new unit tests.
The parser method also includes a bug with quoted values which are not the
last argument. this bug is now fixed too.

Resolves: #72936
Related: #47301
Releases: master, 7.6, 6.2
Change-Id: Ic8df6ea21642e012438dba0a6a299c15939ab119
Reviewed-on: https://review.typo3.org/46428
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>