[TASK] Compatibility to safely deprecate f:case "default" argument
A compatibility layer can be left in place to support "default" used
on f:case. Internally the compatibility ViewHelper changes the class
name used in the ViewHelperNode which gets reflected in the compiled
template as well as the uncompiled rendering.

ReflectionProperty had to be used since no public API exists for
manipulating ViewHelperNodes in this way.

Change-Id: I80c5c8787f6e88f37dfe129ea1844c4421f77d19
Resolves: #73068
Releases: master
Reviewed-on: https://review.typo3.org/46431
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Tested-by: Xavier Perseguers <xavier@typo3.org>