[TASK] Add logic to include languageLablelFiles via return structure
Adds a structure to accept language files like EXT:foo/bar.xlf
in FormEngine elements that are added in inline ajax and on
page load.

Since it is very hard for FormEngine element developers to
handle this case without bugs in own code and renders label
handling useless otherwise, this patch is requested for 7.6, too.

Resolves: #73072
Releases: master, 7.6
Change-Id: Icbfe707000931037b9579bb8bc6a96dfca023d3e
Reviewed-on: https://review.typo3.org/46434
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>