[BUGFIX] Off-by-one error in FAL's LocalDriver
Currently, the LocalDriver of FAL has an off-by-one issue. The passed
starting pointer gets decremented by one, causing issues in the file list
while browsing.

The pointer decrement is removed with this change.

The whole core does not call ``ResourceStorage->getFilesInFolder()``
with the parameters``$start`` and ``$maxNumberOfItems`` having
another value than 0, so this change is assumed to be safe.

Change-Id: I4e24f18b3222f2abdbed83fbbcb18c73d6e52316
Resolves: #73103
Releases: master, 7.6, 6.2
Reviewed-on: https://review.typo3.org/46502
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>