[BUGFIX] impexp ignore file reference records with missing related files
It could happen, that an export contains sys_file_reference records
with relations to sys_files records, that was missing in the exporting
instance. This causes exceptions on importing the sys_file_reference
records, because the related sys_file record is checked on saving of
the sys_file_reference record.
To prevent this this sys_file_reference are ignored on import.

Resolves: #58693
Releases: master, 7.6, 6.2
Change-Id: I68afed93502553b0d55eb858bdb6da5641d1e5f0
Reviewed-on: https://review.typo3.org/46524
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>