[BUGFIX] Drag&Drop sorting respects AuthMode
If an editor is not allowed to edit content elements with a specific
CType (e.g. HTML), drag&drop sorting was still active. Sorting seemed to
work without an error but was blocked in DataHandler due to missing
permissions.
This fix introduces the missing check to deactivate sorting in this case.

Change-Id: Ib40b1e2a2fbbd4694fbb69ed7cda4a2180e15f14
Resolves: #73539
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/46995
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>