[BUGFIX] IRRE child records cannot be synchronized/localized correctly
IRRE child records are not synchronized or localized inside a localized
parent record. The DataHandler command contains the target language that
shall be used for the new localized child elements. Since currently an
array is transmitted, all localized children are created for the first
language (sys_language_uid=1), which is the result of "(int)array(123)".

Resolves: #71596
Releases: master, 7.6
Change-Id: Ic34380c7f773449b62bd1e14318f1ea8c081536f
Reviewed-on: https://review.typo3.org/46819
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>