[TASK] Remove adodb diff
There is a list of patches we applied to adodb manually in
typo3/sysext/adodb/Documentation/Index.rst. This, together
with 'git log' should be enough in case adodb is updated.
The diff file is pain to maintain and also does not
contain all changes that were done to adodb.

Change-Id: If0525ce90b637541d659569f697377f011b8ad37
Resolves: #73607
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/46815
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>