[BUGFIX] Convert colPos in backend layout wizard to number
The method `setColumn` in the backend layout wizard sanitizes the
value by wrapping it into a <p> tag and fetching its text which strips
off any HTML. This returns a string.

However, the code later checks after performing a save action whether
the value is a number. The value gets converted to a number in
`setColumn` to recover the functionality.

Change-Id: I686657e170668adf71f278fa537aa5510e3e2241
Resolves: #73656
Releases: master
Reviewed-on: https://review.typo3.org/46880
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>