[BUGFIX] LocalImageProcessor cannot reuse remote files
The LocalImageProcessor had a check for the storage type, to only do a
check for existing files on local storages, as it needed to fetch an
existing file from the storage to get metadata from it.

This check is now replaced by a proper usage of the FAL API: Fetch the
generated file for local (read-only) processing, which in case of a
local storage means the file is not moved at all. Then, the image
metadata can be extracted from the local file and the file index record
of the processed file can be updated accordingly.

Change-Id: I9fca326fe1d1743cd53a0a85c674ff81e5a051b9
Releases: master, 7.6
Resolves: #67126
Reviewed-on: https://review.typo3.org/47127
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>