[BUGFIX] Set correct foreign parentkey when writing inline relations
When attaching or updating 1:N relations via extbase (INLINE) the
parents uid property was taken as parentKey for the relation entity.

This behaviour is inconsistant to the FormEngine resulting in
wrong weird rendering resconstitution of 1:N relations.

This patch sets the _localizedUid of the parent object (if set)
as parentKey to the parentKeyFieldColumn.

Resolves: #73726
Releases: master
Change-Id: I47fa0f05eb46353f5b893fff07f0123325a0d515
Reviewed-on: https://review.typo3.org/46996
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Timo Pfeffer <pfeffer.timo+typo3@gmail.com>
Reviewed-by: Tobi Kretschmann <tobi@tobishome.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>