[FEATURE] EXT:form - Directly load form wizard inline
This patchset optimizes the form wizard.

The wizard is loaded directly as inline wizard. There is no need anymore
to save and reload the newly created content element in order to be able
to open the wizard. This is a huge usability improvement.
Additionally there is no need to provide individual doc headers. Instead,
the centralized doc headers can be used.

The whole integration utilizes the nodeRegistry of formEngine and
registers the wizard as new render type.

Furthermore, all JavaScript is loaded via require.js.

Since integrators and editors had massive problems with overriden form
configuration the wizard cannot be deactivated anymore. Instead, the
user is globally able to configure the usage of the wizard within the
user settings.

Since it is no longer used we also drop the class WizardView.

Resolves: #69394
Releases: master
Change-Id: I571461b8abf50060791b2f32833d99488eaddaf4
Reviewed-on: https://review.typo3.org/43324
Reviewed-by: Bjoern Jacob <bjoern.jacob@tritum.de>
Tested-by: Bjoern Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>