[BUGFIX] Ensure t3d compatibility for supported TYPO3 version
This patch fixes an issue with unclean t3d export due to string to array
conversion. This is important to be able to import a t3d files which was
exported with 7.6 and above even in 6.2. This is needed as we do not
have any chance to see which TYPO3 version was used for the export.

Resolves: #74127
Releases: master, 7.6, 6.2
Change-Id: I6ba7b825241c2ca439c485aaf597b019b7ac8997
Reviewed-on: https://review.typo3.org/47091
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>