[TASK] Remove fileinfo as dependency in SystemEnvironment/Check
Currenty 'fileinfo' is listed as required php extension in
SystemEnvironment/Check but it is currently not really required
and prevents from installing TYPO3 on Systems where it is not
present (e.g. Mircosoft Azure Cloud)

The class is currently only used once in the core in
"TYPO3\CMS\Core\Type\File\FileInfo" and only when it exists,
so it is not really a hard dependency.

Also grepping on an installation folder does not find any match
where \finfo is really required.

Resolves: #74177
Releases: master, 7.6
Change-Id: I125cd3e8d0eb6f208641375ea71b89917284f8b3
Reviewed-on: https://review.typo3.org/47069
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>