[TASK] Use JsConfirmation enumeration in BackendUserAuthentication
This makes use of the JsConfirmation enumeration to ensure valid values
in BackendUserAuthentication::jsConfirmation.

As combinations of values are allowed, we need to check for all possible
combinations by comparing to an OR-ed bitmask with all possible bits.
Additionally, this fixes a bug where setting the TSconfig option to "0"
would yield a value of 255, which would enable all notifications instead
of disabling them.

Resolves: #69827
Releases: master
Change-Id: I59b8b8de3efb5c33273ea4c7dfdad2ac25473463
Reviewed-on: https://review.typo3.org/43330
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>