[TASK] EXT:form - prevent initial ajax call
Initially, provide the whole configuration for the wizard through a
requireJs config. Therefore, an early AJAX call to the WizardController
can be omitted.

With this patchset, no AJAX route for loading has to be registered. The
loadAction is removed as well. Furthermore, a left over intialize script
is dropped. The Wizard.Settings to configure the ExtJs viewport are used
directly.

Resolves: #74316
Releases: master
Change-Id: I5278784cadc603238707b6e7b26426155e3c080c
Reviewed-on: https://review.typo3.org/47110
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Bjoern Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>