[TASK] Decode body for alternative request methods
PHP only decodes data for POST requests. This commit adds a similar
support (using the same mechanisms, according to the PHP docs) for other
request methods. Currently, it is limited to PUT, PATCH and DELETE, the
request methods that may have a body according to the HTTP spec.

Change-Id: Icdf14f48bf84258958dbca704dfa94474f6894bb
Resolves: #74532
Releases: master
Reviewed-on: https://review.typo3.org/47161
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>