[BUGFIX] Re-allow t3d export from ID 0
During the refactoring of the impexp module a check for the
page ID was introduced to show the pagetree and additional
fields only if a page ID was given, which disallows to export
records on ID 0.

Additionally, a minor JavaScript bug which popped up then
is fixed as well.

Resolves: #75183
Releases: master, 7.6
Change-Id: I0adbd933ecc5da938b511d531178ac707cb035c3
Reviewed-on: https://review.typo3.org/47358
Reviewed-by: Bernhard Kraft <kraftb@think-open.at>
Tested-by: Bernhard Kraft <kraftb@think-open.at>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>