[TASK] ResourceStorage->addFile() should not delete orig file if requested
FALs ResourceStorage->addFile() is the main method to populate a locale
file into a FAL storage. Internally, the method calls the assigned and
protected driver of this storage and calls addFile() on the
driver after checking permissions.
The method *always* deletes the local original file after it has been
added to the storage. This is funny, since addFile() should then call
moveFile(), but moveFile() exists and is only thought to move
files between storages and not from non-storage to storage.
The drivers addFile() however already contains an argument that
suppresses the deletion of the original file, but that argument is
not added to addFile() of the storage ...
The patch now adds this argument to the method signature of
ResourceStorage addFile() to mitigate the issue in a fully
backwards compatible way.
Since the patch is non-destructive and low-risk, the
patch is requested for 7.6, too.

Change-Id: I201c7eac10f7d35db5bdff889b3561ad5b1dce8b
Resolves: #70012
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/47422
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>