[BUGFIX] Prevent warning during extension installation
Currently a warning is triggered if you try to install an extension
which has dependencies as the JsonView uses some own implementation to
convert objects into json with using Reflections. This patch adds some
view configuration to exclude a getter function which needs an argument
to prevent the warning.

Resolves: #75256
Releases: master
Change-Id: Ic63513d36f98f37ae0ab344349d7a85010e5aada
Reviewed-on: https://review.typo3.org/47382
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>