[TASK] Upgrade to phpunit 5.2
With minimum php7 and after codeception released a patch level
that introduced phpunit 5.x support, we can upgrade our phpunit
dev dependency to 5.2 along with a patch level upgrade of vfsStream.

This phpunit version is a bit more strict, some unclean tests
stumble and need fixes:

* $this->setExpectedException() needs a string as second argument,
all places with $this->anything() or $this->any() to ignore the
actual exception string are now substituted with empty string, the
tests rely on the timestamp to specify a specific exception.

* $aMock->expects()->method('aMethod') now throws exceptions if a
method name is specified that is either not mocked in $this->getMock()
or that does not exist in the mocked class. This was previously
ignored. This change now catches a lot of code smell in our tests :)

Change-Id: I546f8f4b3e05b3d643f4ee6be8e1d4c6e23fe083
Resolves: #75264
Releases: master
Reviewed-on: https://review.typo3.org/47388
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>