[BUGFIX] Treat numbers for comparison in FormEngineValidation as integer
This patch fixes the issue that the number-based rules are treated
as strings. The numbers are converted to integers now.

Also, the ``range`` is changed: Instead of defining any fallback
values if either ``lower`` or ``upper`` are not set, the value is now
only compared against those limits when they exist.

Resolves: #75307
Releases: master, 7.6
Change-Id: Iab321a7637e2ccc8c662355ccef1e25bc0b78be7
Reviewed-on: https://review.typo3.org/47424
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>