[TASK] Add index tags to documentation rst files
As a service to the user searching for a specific topic in changes,
documentation files receive tags to be used as index entries.
The rendering still needs to be adjusted to display this index, this
will be done in a seperate patch. For now, only the tagging has been
done.

Releases: master
Resolves: #75382
Change-Id: If38af8ff989954882e9af51625e89d51671e4c66
Reviewed-on: https://review.typo3.org/47499
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>