[TASK] Clean up unit tests on Windows
At the moment a lot of tests are skipped on Windows environments.
The main reasons are permissions, symlinks and shell exec commands.
As symlinks are now working on Windows (if the user / group is allowed
to create them) there is no reason to skip those tests. As you can only
create valid symlinks on Windows (the target is checked for existence)
the tests needed to be adjusted to take that into account.

Additionally some of the other tests don't need to be skipped either, as
they mock all system specific calls (e.g. AbstractNodeTests).

For the other tests the message on skipping was updated to contain more
information as to what causes them to be irrelevant on Windows.

Resolves: #75388
Releases: master
Change-Id: I86c6d9b0cec34a94d1185f46e2023af811b16107
Reviewed-on: https://review.typo3.org/47502
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>