[SECURITY] Disallow login with empty password
In case a backend or frontend user is stored in the database
with an empty string as password (not possible through backend UI),
it is possible to authenticate this user using an empty password
with the standard TYPO3 username/password authentication services.

By definition this should be prohibited.

Resolves: #75055
Releases: master, 7.6, 6.2
Security-Commit: 1899bfa7166baae8d774fa7bd027f9c448e89686
Security-Bulletins: TYPO3-CORE-SA-2016-009, 010, 011, 012
Change-Id: I7b5ce35a6e5d817c2480cb81e616bfac25fbe2fb
Reviewed-on: https://review.typo3.org/47599
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>