[SECURITY] Disallow login with empty password
In case a backend or frontend user is stored in the database
with an empty string as password (not possible through backend UI),
it is possible to authenticate this user using an empty password
with the standard TYPO3 username/password authentication services.

By definition this should be prohibited.

Resolves: #75055
Releases: master, 7.6, 6.2
Security-Commit: b8e1cf8c771e9908c2ab7552b0f9a1c566365879
Security-Bulletins: TYPO3-CORE-SA-2016-009, 010, 011, 012
Change-Id: I0ac861d44fe5a3b110c2f5de7ef19b458cec2c79
Reviewed-on: https://review.typo3.org/47603
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>