[TASK] Provide QueryHelper to prepare SQL fragments for the QueryBuilder
The QueryHelper deals with GROUP BY and ORDER BY strings coming from
sources where the data format is not strictly defined and the input
might be a string prefixes or otherwise unsuitable for immediate further
usage.

For GROUP BY strings the prefix is removed and the remainder of the
string is treated as a comma separated list of column names that get
parsed into an array.

For ORDER BY strings the prefix is removed and the remainder of the
string is treated as a list of comma separated column names, possibly
with a sort direction separated from the column name, that gets parsed
into an array.

For WHERE SQL fragments leading the leading logical operators "AND" and
"OR" get removed, no further modifications are performed on the string.

Resolves: #75559
Releases: master
Change-Id: I37016eda6597cab49b274f498188a8579dec7b2c
Reviewed-on: https://review.typo3.org/47636
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>